ethiomart_amharic_ner_project/requirements.txt
transformers>=4.30.0
datasets>=2.12.0
torch>=1.13.0
telethon>=1.27.0
pandas>=1.5.0
scikit-learn>=1.1.0
seqeval>=1.2.2
shap>=0.41.0
lime>=0.2.0.1
matplotlib>=3.5.0
python-dotenv>=1.0.0
tqdm>=4.64.0
